+++ src/frontend/magnify/src/controller/mock.controller.ts
@@ -3,4 +3,6 @@ import Controller from './interface';
 export default class MockController extends Controller {
   sendTest = jest.fn();
   getExamples = jest.fn();
+
+  joinRoom = jest.fn();
 }